rename memory to cache
