+++ pkg/cache/simple/simple_test.go
@@ -1,4 +1,4 @@
-package memory
+package simple
 
 import (
 	"fmt"
